Refactor some code into more files
